add mutations for users crud

+++ crud-backend/src/graphql/user.resolver.js
@@ -1,7 +1,8 @@
+const boom = require("@hapi/boom");
 const UserService = require("../services/user.service");
 const service = new UserService();
 
-const getUser =  (_, { id }) => {
+const getUser = (_, { id }) => {
   return service.findOne(id);
 };
 
@@ -9,18 +10,29 @@ const getUsers = () => {
   return service.find({});
 };
 
-const addUser =  (_, {dto}) => {
+const addUser = async (_, { dto }, context) => {
+  const { user } = await context.authenticate("jwt", { session: false });
+  if (!user) {
+    throw boom.unauthorized("You must be logged in");
+  }
   return service.create(dto);
-
 };
 
-const updateUser = (_, {id,dto}) => {
+const updateUser = async (_, { id, dto }, context) => {
+  const { user } = await context.authenticate("jwt", { session: false });
+  if (!user) {
+    throw boom.unauthorized("You must be logged in");
+  }
   return service.update(id, dto);
 };
 
-const deleteUser = async (_,{id}) => {
-  const user = await service.delete(id);
-  return user;
+const deleteUser = async (_, { id }, context) => {
+  const { user } = await context.authenticate("jwt", { session: false });
+  if (!user) {
+    throw boom.unauthorized("You must be logged in");
+  }
+  const deleteData = await service.delete(id);
+  return deleteData.id;
 };
 
 module.exports = { getUser, getUsers, addUser, updateUser, deleteUser };
